+++ trunk/TAO/TAO_IDL/be/be_visitor_argument/invoke_cs.cpp
@@ -645,7 +645,7 @@ int be_visitor_args_invoke_cs::visit_string (be_string *node)
             case AST_Argument::dir_IN:
               break;
             case AST_Argument::dir_INOUT:
-              if (node->width () == sizeof (char))
+              if (node->width () == (long) sizeof (char))
                 {
                   *os << "CORBA::Any::to_string (";
                 }
@@ -658,7 +658,7 @@ int be_visitor_args_invoke_cs::visit_string (be_string *node)
                   << node->max_size ()->ev ()->u.ulval << ")";
               break;
             case AST_Argument::dir_OUT:
-              if (node->width () == sizeof (char))
+              if (node->width () == (long) sizeof (char))
                 {
                   *os << "CORBA::Any::to_string (";
                 }
